Move styles for custom css
