+++ src/components/custom-css-control/editor.scss
@@ -1,0 +1,56 @@
+.maxi-additional__css {
+	.maxi-base-control {
+		&__field {
+			align-items: start;
+			flex-direction: column;
+		}
+	}
+	.maxi-base-control__field {
+		display: block !important;
+		.maxi-base-control__label {
+			max-width: 70%;
+			margin-top: 7px;
+		}
+		.components-button {
+			float: right;
+			height: 24px;
+			background-color: var(--maxi-grey-9-color);
+			border: 1px solid var(--maxi-grey-1-color);
+			border: 1px solid var(--maxi-grey-1-color);
+			border-radius: 0;
+		}
+	}
+	.w-tc-editor {
+		margin-top: 8px;
+		width: 100%;
+		background-color: var(--maxi-white-4-color);
+		border: 1px solid var(--maxi-grey-1-color);
+		font-family: monospace !important;
+
+		textarea {
+			background-color: var(--maxi-white-4-color);
+			font-family: monospace !important;
+		}
+	}
+	code {
+		padding: 0;
+		background-color: var(--maxi-white-4-color);
+		font-family: monospace !important;
+		span {
+			background-color: var(--maxi-white-4-color);
+			font-family: monospace !important;
+		}
+	}
+	.maxi-additional__css-error {
+		margin-top: 8px;
+		padding-left: 10px;
+		padding-right: 5px;
+		font-style: italic;
+		&.not-valid {
+			border-left: 2px solid var(--maxi-primary-color);
+		}
+		&.valid {
+			border-left: 2px solid var(--maxi-green-1-color);
+		}
+	}
+}
