+++ src/components/index.js
@@ -20,7 +20,7 @@ export { default as CheckBoxControl } from './checkbox-control';
 export { default as ClipPath } from './clip-path-control';
 export { default as ColorControl } from './color-control';
 export { default as ColumnPattern } from './column-pattern';
-export { default as CustomCss } from './custom-css';
+export { default as CustomCss } from './custom-css-control';
 export { default as CustomLabel } from './custom-label';
 export { default as DefaultStylesControl } from './default-styles-control';
 export { default as DisplayControl } from './display-control';
